requirements.txt
altgraph==0.17.4
Babel==2.15.0
bidict==0.23.1
blinker==1.8.2
boto3==1.35.87
botocore==1.35.87
chromedriver-autoinstaller==0.6.4
click==8.1.7
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
dnspython==2.7.0
et-xmlfile==1.1.0
eventlet==0.37.0
Flask==3.0.3
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
numpy==1.26.4
openpyxl==3.1.3
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.4.0
psycopg2-binary==2.9.10
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.1
pywin32-ctypes==0.2.2
s3transfer==0.10.4
setuptools==71.1.0
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.36
tkcalendar==1.6.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
Werkzeug==3.1.2
wsproto==1.2.0
XlsxWriter==3.2.0
